requirements.txt
agno>=0.1.0
python-dotenv>=1.0.0
plotly>=5.18.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
bs4>=0.0.1
beautifulsoup4>=4.12.0
yfinance>=0.2.35
geopy>=2.4.1